added auto start x

+++ .bashrc
@@ -1,6 +1,12 @@
+# bashrc
+
+# If not running interactively, don't do anything
+if [[ $- != *i* ]]; then return; fi
+
 PS1='\w\$ '
 
 # alias
+#alias ls='ls --color=auto'
 alias ll='ls -alF'
 alias rm='rm -rf'
